+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fixed `SelectionList` issues after removing an option https://github.com/Textualize/textual/pull/4464
+- Fixed `ListView` bugs with the initial index https://github.com/Textualize/textual/pull/4452
+
+### Changed
+
+- When displaying a message using `App.exit()`, the console no longer highlights things such as numbers.
 
 ## [0.58.1] - 2024-05-01
 
@@ -18,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed issue with Markdown mounting content lazily https://github.com/Textualize/textual/pull/4466
 - Fixed intermittent issue with scrolling to focus https://github.com/Textualize/textual/commit/567caf8acb196260adf6a0a6250e3ff5093056d0
 - Fixed issue with scrolling to center https://github.com/Textualize/textual/pull/4469
+
 
 ## [0.58.0] - 2024-04-25